% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyf.R
\name{ganancia}
\alias{ganancia}
\title{Calcula ganancia de la prediccion}
\usage{
ganancia(
  data = NULL,
  real = NULL,
  prediccion = NULL,
  imprimir = T,
  devolver = F
)
}
\arguments{
\item{data}{dataframe con las columas 'real' y 'prediccion'}

\item{real}{vector con los valores reales de los BAJA+2}

\item{prediccion}{vector con los valores predichos de los BAJA+2}

\item{imprimir}{T = imprime por pantalla, F = no imprime}

\item{devolver}{T = return ganancia, F = no devuelve nada}
}
\description{
Calcula ganancia de la prediccion
}
