% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyf.R
\name{kaggle_csv}
\alias{kaggle_csv}
\title{Exporto csv para subir a kaggle}
\usage{
kaggle_csv(
  data = NULL,
  clientes = NULL,
  estimulos = NULL,
  path = "~/input_kaggle.csv"
)
}
\arguments{
\item{data}{dataframe con las columas 'numero_de_cliente' y 'estimulos'}

\item{clientes}{vector con los numeros de clientes}

\item{estimulos}{vector con los estimulos de cada cliente}

\item{path}{path del archivo de salida. Si existe, se crea un 'input_kaggle_YYYYMMDD_HHmmss.csv' por default, en la carpeta del usuario}
}
\description{
Exporto csv para subir a kaggle
}
