% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyf.R
\name{levantar_clientes}
\alias{levantar_clientes}
\title{Levanta dataset y arma clase binaria}
\usage{
levantar_clientes(
  path,
  nombre_clase_binaria = "baja",
  positivo = "si",
  negativo = "no",
  fix_nulos = F
)
}
\arguments{
\item{path}{path donde esta el csv con los clientes}

\item{positivo}{valor positivo -factor-}

\item{negativo}{valor negativo -factor-}

\item{fix_nulos}{corregir nulos}

\item{dataset}{dataset a modificar -TIENE QUE SER EL DE LA MATERIA-}

\item{nombre}{nombre de la clase binaria}
}
\description{
Levanta dataset y arma clase binaria
}
