% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyf.R
\name{rf_ranger}
\alias{rf_ranger}
\title{Train y test random forest usando 'ranger'}
\usage{
rf_ranger(
  a_predecir,
  positivo,
  n_arboles,
  n_split,
  nodo_min,
  profundidad_max,
  entrenamiento,
  evaluacion,
  salida_csv,
  ver_ganancia = T,
  kagglecsv = T
)
}
\arguments{
\item{a_predecir}{string con la variable a predecir -en función del resto de las variables-}

\item{positivo}{string con el valor positivo de la variable a predecir}

\item{n_arboles}{numero de árboles}

\item{n_split}{numero de variables posible a dividir en cada nodo}

\item{nodo_min}{tamaño minimo del nodo}

\item{profundidad_max}{profundidad máxima de los árboles}

\item{entrenamiento}{dataset de entrenamiento}

\item{evaluacion}{dataset de evaluación}
}
\description{
Train y test random forest usando 'ranger'
}
