% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyf.R
\name{rt_rpart}
\alias{rt_rpart}
\title{Train y test árbol de regresión usando 'rpart'}
\usage{
rt_rpart(
  a_predecir,
  positivo,
  complejidad,
  n_split,
  nodo_min,
  profundidad_max,
  entrenamiento,
  evaluacion,
  n_cv = 0,
  salida_csv = "~/rt_rpart.csv",
  ver_ganancia = T,
  kagglecsv = T
)
}
\arguments{
\item{a_predecir}{string con la variable a predecir -en función del resto de las variables-}

\item{positivo}{string con el valor positivo de la variable a predecir}

\item{complejidad}{complejidad utilizada para partir los nodos}

\item{n_split}{numero de variables posible a dividir en cada nodo}

\item{nodo_min}{tamaño minimo del nodo}

\item{profundidad_max}{profundidad máxima de los árboles}

\item{entrenamiento}{dataset de entrenamiento}

\item{evaluacion}{dataset de evaluación}

\item{n_cv}{número de veces a realizar cross validation}
}
\description{
Train y test árbol de regresión usando 'rpart'
}
