% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyf.R
\name{feature_eng}
\alias{feature_eng}
\title{Ingeniería de features para el dataset crudo de clientes.}
\usage{
feature_eng(
  d,
  historico_desde = 202001,
  top_variables_importantes = 10,
  historico_de = 7,
  combinar_tarjetas = T,
  historicos_de_tarjetas = T,
  borrar_originales_de_tarjetas = F,
  correccion_catedra = T,
  devolver_modelo = F
)
}
\arguments{
\item{d}{dataset clientes original -se devuelve copia-}

\item{top_variables_importantes}{número de variables importantes a tener en cuenta para el cálculo de históricos}

\item{historico_de}{número de meses anteriores que usa para calcular el histórico del mes actual}

\item{desde}{mes desde el cual se empieza el procesamiento -luego, se procesan los meses anteriores-}
}
\description{
Ingeniería de features para el dataset crudo de clientes.
}
